requirements.txt
redis
click
gunicorn==19.9.0
google-cloud-logger==0.2.1
google-cloud-core==1.0.3
google-cloud-logging==1.14.0
google-cloud-pubsub==0.40.0
google-cloud-storage==1.20.0
google-cloud-monitoring==1.0.0
pulsar-client==2.4.2
bcrypt==3.1.3
flask==1.0.0
flask-wtf==0.14.2
flask-cors==3.0.7
flask-marshmallow==0.8.0
flask-restful==0.3.7
flask-session
fastavro
minio==5.0.6
Werkzeug==0.16.0
